Create UserEventSyncBridge.
This is the first step towards implementing UserEvents and integrating
with Sync. This CL adds a ModelTypeSyncBridge implementation, but it is
never instantiated or called, yet.

While adding unit tests for UserEventSyncBridge, I added
ModelTypeSyncBridge::ChangeProcessorFactory which was usable in both
auto complete and device info sync bridge unit tests.

BUG=701032

Review-Url: https://codereview.chromium.org/2856933005
Cr-Commit-Position: refs/heads/master@{#469829}