Add consent_specifics proto.
Add consent_specifics proto as a preparation for separation of consents
into a separate model type from user events. The proto is based on
UserEventSpecifics::UserConsent. The proto has to land before the actual
split itself, so that the server is aware of the new datatype.

Bug: 840357
Change-Id: I2d3ab6575990d081315a93725fb5cee655af63b9
Reviewed-on: https://chromium-review.googlesource.com/1055567
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Markus Heintz <markusheintz@chromium.org>
Commit-Queue: vitaliii <vitaliii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#559453}