Support scrypt key derivation under feature
Add scrypt as a key derivation method to the proto and also to the
client-side KeyDerivationMethod enum. Add feature which, when disabled,
effectively treats scrypt as an unknown key derivation method
altogether.

Wire up scrypt key derivation in Nigori, so that it is exposed
externally.

Add tests to ensure that Nigori correctly derives scrypt keys, that
SyncEncryptionHandlerImpl passes the new key derivation method correctly
to observers and that, when the feature is disabled, scrypt code paths
are not triggered. This is beneficial in mitigating risk related to the
introduction of scrypt, as it can be completely disabled via Finch if
need be.

Bug: 877933
Change-Id: Icf940a02ef03b172c84d828d96b264c574151f7c
Reviewed-on: https://chromium-review.googlesource.com/1194006
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: vitaliii <vitaliii@chromium.org>
Commit-Queue: David Davidović <davidovic@google.com>
Cr-Commit-Position: refs/heads/master@{#587689}