Wrap channels one-of into message.
This way it can be later extended easier (e.g. tag number space is
separate from outer SharingMessageSpecifics) and also passed as an
argument in C++ code.

This is the first step to start building SharingMessageSpecifics in
the bridge instead of doing this in Sharing code.

Changing tag numbers in the proto is fine, because
a) it is not used anywhere yet
b) it is not persisted anywhere anyway.

Bug: 1044537
Change-Id: I4201680cfd8149888efbc35da6fbb5a7969f436a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013246
Auto-Submit: vitaliii <vitaliii@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733965}