Fetch bookmark final GUIDs upon startup if safe
This is a soft version of the logic introduced with crrev.com/c/2278250,
still disabled by default. The new logic is considered harmless and
hence enabled by default.

The ultimate goal continues to be to simplify the logic of dealing with
sync-ed bookmarks by adopting the invariant that all final GUIDs are
known (which is the same as knowing client tags).

Making the validation of sync metadata more strict upon startup has the
downside that sync will go through the initial sync procedure
(BookmarkModelMerger) upon startup and that incurs in some risk for the
user, for example if a pending local tombstone exists (i.e. the user
deleted a bookmark but sync did not manage to upload this change).

In order to minimize this risk, this patch introduces heuristics that
determine, upon startup, whether the local client is in sync with the
server. If yes, then the desired invariant (all final GUIDs known) is
enforced, which means local sync metadata may be thrown away and
initial sync exercised. Otherwise, nothing happens and the logic may
kick in upon the next statup, when prerequisites get reevaluated.

In this variant, in combination with the numbers gathered by the UMA
metric Sync.BookmarkModelMetadataClientTagState, the chances for a
user to run into undesired behavior (e.g. undeleted bookmark) should
be negligible.

Bug: 1032052
Change-Id: I21234ca2a93e080e2c81bfdf43b915150e5bdfb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296420
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788149}