Add interested data types to device info specifics.
This CL adds a field to device info specifics for a list of data types for which the device wants to receive invalidations.
The list will be used by the sync server to send invalidations for only those types.

The data types are obtained from SyncInvalidationsService.

Bug: 1102312
Change-Id: Ia85f2b523d1f9adeef84dc4562d1bd525615e368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2387113
Reviewed-by: Jeffrey Cohen <jeffreycohen@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Paula Vidas <paulavidas@google.com>
Cr-Commit-Position: refs/heads/master@{#804144}