Introduce an FCM token list as part of a commit message
The list can be used by the server to send invalidations to other
clients on commit request. The list contains all known FCM tokens from
all active devices known to the client via DeviceInfo data type.

A device is considered active if it sent DeviceInfo specifics within its
pulse time interval (with an additional margin interval).

The list might contain the FCM token of the local device to support
reflections. There is also a limit of the size of the list to restrict
the additional data in a commit request.

The empty list means that there is no known information about all other
clients and the server should decide itself where to send invalidations.
To distinguish the case when there is the only client (which doesn't
expect a reflection invalidation), the client should provide the single
client flag.

Bug: 1164849
Change-Id: I4770091d8e5a577ad9581f899676feeaf381ba97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2620560
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/master@{#843003}