Delete some unused proto fields in CommitResponse
These fields are literally dead code on the client since M86, when the
pre-USS codebase was deleted, but older versions of the browser
(manually verified with M56) considered them optional and barely used
them anyway, because it was only used within GetResultingPostCommitName() where it fell back to
the name of the entity being committed (i.e. in case the server
overwrote the name during commit, which doesn't happen in practice).

Note also that the server doesn't populate these fields, except for
|name| which is indeed populated for legacy reasons.

Change-Id: I764c0e294d46a8a090b850a6114b4894a8708aba
Bug: 1182252
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2720142
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858571}