Add tabbed as a new display mode.
This change adds a new display mode "tabbed" which is only valid when
used as a display_override.

Tested:
- PWA with display:"standalone", display_override:["tabbed"] is
installable and tabbed is a valid display override when flag is enabled.
- PWA with display:"standalone", display_override:["tabbed"] is
installable but tabbed is not a valid display override when flag is
disabled.
- PWA with display:"tabbed" is not installable.

Bug: 897314
Change-Id: I140764f332a9318597d59adbbd277b7158b23738
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2929540
Reviewed-by: Markus Heintz <markusheintz@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Louise Brett <loubrett@google.com>
Cr-Commit-Position: refs/heads/master@{#891499}