Expose a bookmark's parent's GUID in sync protocol (specifics)
The field is somewhat redundant to the pre-existing parent_id proto
field, which is server-defined. The new field (a GUID in specifics)
has a number of advantages:
1) Is very simple, and in particular does not require fancy dealing with
   temporary IDs during commits.
2) Is more robust agaist hash collisions and server-side proto
   serialization bugs.
3) Is better privacy-wise, since it can be encrypted (for custom
   passphrase users).
4) Unblocks better dealing with orphan nodes, which would now at least
   determine the parent's GUID, and hence such a folder can in fact be
   created (with a proper GUID) without having any further information.

That said, none of the advantages listed above will be achieved for real
anytime soon, since old data will continue to exist for a long while.
Nevertheless, let's set the foundation for a better long-term future.

Change-Id: Icfe2530a679b519e54115a378d9a67a6e7f239d2
Bug: 1143246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3053690
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905361}