Add result proto for list calls to Chrome Sync module
Requests issues to the ChromeSync module in Google Mobile Services
respond to list requests with the newly introduced ListPasswordsResult
proto.
Right now, it only wraps PasswordData but will contain additional meta
data in the future. The PasswordData consists of a synced component that
is identical to synced passwords across all platforms and an unsynced
BLOB that is kept local and is managed only by Chrome on Android.

Although Chrome keeps the source of truth, the protos will be synced for
internal use.

Bug: 1229654
Change-Id: Icf62b184057ec5572d84df2377ebaae940f09fcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3081823
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910241}