Omit bookmark favicon URL if too large
Occasionally, a bookmark may have an associated favicon with an URL that
is too large to sync (several kilobytes), preventing the whole bookmark
from sync-ing successfully, because of the server-enforced maximum size
of a sync-ing entity (bookmark).

In practice, this case has been reported by users for pages that encode
the content of the favicon image itself in the favicon URL, by prefixing
the URL with the data: scheme, which means the URL can grow arbitrarily
large.

There is no good reason to stop sync-ing a bookmark because of the
favicon's URL, so this patch simply changes the upload (commit) codepath
to omit the field altogether if too large. This is fine because there is
logic in place already for handling the favicon-URL-missing case, used
prior to this patch for other scenarios.

The logic is introduced with a kill switch, in case it runs into issues,
but otherwise enabled by default.

Bug: 1252119
Co-Authored-By: victorvianna@google.com
Change-Id: I7e733dba3d9cad4e8f546f60c82048aa88cb6c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3535786
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#982730}