Sync password notes only when notes feature enabled
Before this CL:
Chrome will unconditionally sync password notes to the sync server.
When the notes features is disabled, the client doesn't have a reliable
value of the note, and hence shouldn't send it to the server.

After this CL:
Chrome will only sync password notes to the sync server when the
notes feature is enabled.

This CL is also adding the proper testing that confirms that proto
conversion works as expected with and without the notes feature.

Since the code that unconditionally writes notes has been rolled out
to public clients already, this CL is deprecating the proto fields
that used to carry notes and notes backup and introduces new fields
instead.

Although the notes backup field is gated behind a feature toggle, it
is decided to also deprecate it for future simplicity and ease
of debugging.

The currently implemented server workaround will *not* be effective
until the server picks-up the proto changes.

LOW_COVERAGE_REASON=predates this patch, and integration test will be added in a follow up.

Bug: 1368988
Change-Id: Ic3d320c277a71ff31279d38feb06aa0d6e2745aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925585
Auto-Submit: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1054013}