Disable sync if remote bookmarks exceed limit in initial merge.
Bookmarks' merge operation happens in the UI thread. When there's a
large number of bookmarks to be synced, the long processing freezes the
UI thread, which causes problems. This change is part of a set of CLs to
limit the number of synced bookmarks. This CL disables sync if count of
remote updates exceed limit. The count is saved in the sync metadata to
prevent re-download of bookmarks. To re-enable sync, user would have to
turn off and turn on sync.

Bug: 1347466
Change-Id: Ib7d4938531b7bd46053214a31886409b919c9802
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4006654
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1070351}