Include the uncompressed largeBlob size in the WebAuthn credential.
The blobs are DEFLATE compressed, per the spec, before being written to
security keys to make the most of the limited storage space. We'll do
the same for phones too, thus we also need to store the uncompressed
size.

There's no visitor updates in this change because the new field is in
the encrypted submessage which the visitors see only via the encrypted
bytes field.

Change-Id: I09ecc0bb0e79400c324c5ec2ccc31056fff11866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4179387
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1095566}