Rename SyncEntity.client_defined_unique_tag
The previous name generated confusion because it doesn't convey the
need to hash the value. While in theory the protocol can support
unhashed tags, in practice this scenario is extremely unlikely
considering the client-side architecture.

To avoid mis-use, this patch renames the field to adhere to more
common terminology and a less ambiguous one too.

Change-Id: I89b4b60a29c5249fdb3f13c3236285489a055b20
Fixed: 1422202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4308278
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1115052}