webauthn: add `key_version` to entity protobuf.
The security domain key used to encrypt the passkey secrets can change.
In order to avoid doing trial decryption when several key are possible,
this change adds a field to the protobuf that specifies which key (by
version) was used to encrypt the included secrets.

Bug: b:290075051
Change-Id: Iee5eecbd03a8eb6efe18fb886b27c286b162f196
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4684446
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171821}