Add DeviceInfo field for the Google sign-in credential ID.
Since this key is autogenerated, it's not published as a
WebauthnCredentialSpecifics like other GPM passkeys. So, in order to
make various features work on accounts.google.com as well as they do on
other sites, this change allows a phone to publish a credential ID for
the autogenerated key.

Bug: b:260652806
Change-Id: I7afd06a2f7219307c9c96f53bc38752ebc8a6bc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4718493
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1175509}