Add promise_package_id to sync data
Add promise_package_id field to sync data. When the sync service
triggers an app installation on a device, we will create a promise app
item to represent the installation in the Launcher/ Shelf. We need to
check the sync data for the correct ordinal positions so that we can
place the item in the position in the Launcher/ Shelf. To know which
sync item we should read, we need a common ID between the promise app
item and the sync item. This common ID is the promise_package_id
field, which is the ID of an app package that is globally unique
across platforms. The Promise Package ID of an app is consistent
whether the app is installed or not. We cannot use the existing sync
App ID field for matching, since it can only be generated after an app
is installed (so a promise app item is incapable of forming one).

Bug: b/261907612
Change-Id: I3cbbaab2ef19cc531314efedf15cc63bd7aa92bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4677221
Commit-Queue: Vivian Pao <vpao@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181848}