Sync ProfileTokenQuality observations.
For every stored type, AutofillProfile::token_quality() stores a set of
up to 10 observations. An observation consist of an ObservationType,
which is the value of a small enum, and a 1 byte hash used for some
deduplication.
This CL adds logic to sync these observations for account profiles. For
local-or-syncable profiles, observations won't be synced. For details,
see [1].

Technically, this CL:
- Adds a repeated Observation field to the
  ContactInfoSpecifics::TokenMetadata proto.
- Modifies the AutofillProfile <-> ContactInfoSpecifics conversion logic
  to populate the new field and convert back to AutofillProfile.

To keep the hash private, the conversion logic is made a friend of
ProfileTokenQuality.

To support future ProfileTokenQuality::ObservationTypes, the observation
types are stored as ints in the proto. Since the C++ side enum is an
enum class, unknown values will be preserved.

[1] https://docs.google.com/document/d/1-hoBeCogz8gewu3rfhRNGTWObqWfwSSNZ29ikYGml2I/edit?resourcekey=0-vK0gJwbg5cmepJJHBGCC7g

Bug: 1453650
Change-Id: Ic9b47696651b92d453b34e2253f99d9aff31d214
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4759008
Reviewed-by: Marc Treib <treib@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1189412}