Proto field for shared password sender profile image url
This CL is only introducing proto field. Follow-up CLs will populate
and read the field in the regular password sync flow.

This is split in two CLs to avoid issues on the server when new fields
are populated if they are unknown to passwords.google.com.

Bug: 1448235
Change-Id: I51af76a3c578d7047c8333d7774899e9805cc101
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4952980
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212230}