Whitelist -> allowlist under safe_browsing
We need to rename the fields in a couple of protos. This is safe as long
as these protos are only stored/transmitted in serialized binary form,
as then fields are identified by tag number.

Change-Id: Ia614aceaa5d5d512fa8aa785ad4ba4ba776d4338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5028454
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225714}