Introduce session_start_time in SessionHeader and populate it
The timestamp represents when the session was created, i.e. when
the user turned on Sessions sync on this device.
It's only populated for sessions created after this CL; for
preexisting sessions it'll remain empty.

Bug: 364234413
Change-Id: Ib38d6735dda087ec1ae6bb17d98b8454d84ed532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5830790
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ankush Singh <ankushkush@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351287}