Relax DCHECK when producing SessionSpecifics
In rare cases, navigation URLs may be invalid when serializing tabs to
protobufs (SessionSpecifics), which can lead to hitting a
DUMP_WILL_BE_NOTREACHED() in `GURL::spec()` before this patch. As
argued below, this scenario is harmless, hence this patch proposes to
use empty strings without exercising GURL::spec() if the URL is
invalid during serialization, following the behavior before this patch
in DCHECK-disabled builds.

At a first glance, it seems like `SessionsClient::ShouldSyncURL()`
guarantees that all URLs exposed to sync are valid. However, this
predicate is not strictly enforced to all URLs, and it is generally
enough that at least one URL in a tab's controller history passes the
predicate, for sync to upload all recent URLs in this tab (valid or
not).

Similarly, URLs restored from disk are not filtered using the predicate
alluded to above, so disk corruption issues could also hit the DCHECK/
NOTREACHED before this patch.

Relaxing the DCHECK is just fine, because the receiving end (or
consuming end) of sync is anyway responsible for re-exercising
`SessionsClient::ShouldSyncURL()` before exposing URLs to UI / feature
code. This is already done today as part of
`OpenTabsUIDelegateImpl::GetForeignSessionTabs()`, so invalid URLs can
be considered harmless and the DCHECK/NOTREACHED relaxed as proposed in
this patch, and can be treated as empty strings just like
`GURL::spec()` does in DCHECK-disabled builds.

Change-Id: I4d1109ac9235133daad79e3bd5e61bc931b2aaa5
Fixed: 40939540
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198503
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1411658}