Remove redundant 'state' field in per-extension preferences
The 'state' field is currently used to track the enabled state of
extensions. However, this field is redundant because the
'disable_reasons' field already provides this information. If there are
no disable reasons, the extension is enabled; if there are disable
reasons, the extension is disabled.

Due to some existing bugs, the 'state' and 'disable_reasons' fields can
become out-of-sync, leading to hard-to-debug issues.

This change removes the 'state' field. The enabled state of extensions
will now be determined solely based on the 'disable_reasons' field. This
will eliminate the possibility of these fields going out-of-sync and
reduce the likelihood of related bugs.

Bug: 40554334,372186532

Change-Id: I5b75d5845c34cdbd9fa8d77bcde7ed406534f930
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6295214
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Sohail Rajdev <sorajdev@microsoft.com>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1427960}