Add metadata for storage in web applications system
This CL creates the data structure for the metadata that will be used
for storing trusted icons in the web applications system, as per the
design in go/pwa-primary-icon. This CL adds:
1. A field in web_app.h to store the trusted icons.
2. A field in web_app.proto to store the trusted icons on disk.
3. A field in web_app_specifics.proto to ensure the trusted icons can be synced.
4. A field in web_app_install_info.h to ensure that the icons can be parsed during implementation.
5. Serialization code to convert these fields to be supported by the
protobufs, along with tests that verify storage.
6. Logic that sets all the fields correctly in the web_app.

The ability to apply the synced trusted icons as part of sync installs
will be added in a follow-up CL.

Test flakiness has been skipped due to crbug.com/410653528

Bug: 427566601
Change-Id: I249383b16908b8548a9df02717f6d1076b53dccf
Validate-Test-Flakiness: skip
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6761732
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: May Siem <msiem@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1488339}